placeholder response algorithm pt. 4
The script now has a functioning rule-following placeholder algorithm, acting as scaffolding to the real more complicated algorithm to be built later. Right now, this makes a list of possible moves using the current board/grid position and the last move which directs where the next X/O can be placed.