player selection locked in box corresponding to the relative position of last move
The script now locks the player's selection to the box corresponding to the relative position of the smaller square of the opponent's last move. However, if said box has ended in a win, lose, or draw, the opponent is not locked and can freely select any square in any box.