win and draw bitmaps
There are now bitmaps for X winning, O winning, and a draw that are currently unused but will be placed over a tic tac toe box when it is finished. Each of the symbols are separated into 9 quarters, each to go above it's respective square, minus the empty ones leaving us with: 8 for O, 3 for X, and 3 for T (tie). X and T only have 4 empty tiles/squares, however 2 are copies of the other.