different box highlighting method
Now when the user is selecting a box (not a square) it highlights in a different way compared to before. It used to change each one of the 9 tiles inside of the box to a red version of itself, however now it just changes the borders around the selected box to red.