2 Player + Debuging
The "I" button is fixed and now doesn't select a won/drawn tic tac toe game square. I also decided to change the game from a random response algo as a placeholder for the real algorithm, to a simple 2 player game, as I realized I most likely will not be able to complete an algorithm, however if anyone wants to attempt to finish the response algorithm, just pull last commit and change teh placeholder response algo.